Move VarTableEntry to new file
